jinahub/encoders/text/TextPaddleEncoder/requirements.txt
paddlepaddle==2.1.0
paddlehub==2.1.0
sentencepiece==0.1.96
jina-commons @ git+https://github.com/jina-ai/jina-commons.git@v0.0.1#egg=jina-commons